Remove the traitCollection fallback on iOS 13
On iOS 13 UIKit now predicts the initial traits for a view before it's rendered so no need to return the screen's trait collection instead. Moreover, it's better to use the app's key window for determing the fallback traitCollection instead of the screen since for apps that support split screen that will be different based on the configuration chosen by the user.